chore: refactor phase 3
