+++ internal/git/commit_planner.go
@@ -22,77 +22,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"path/filepath"
-	"time"
-
-	gogit "github.com/go-git/go-git/v5"
-	"sigs.k8s.io/controller-runtime/pkg/log"
 )
-
-// PendingWriteKind distinguishes the durable write shapes retained until push.
-type PendingWriteKind string
-
-const (
-	// PendingWriteGroupedWindow is buffered live-event work drained from the
-	// commit window.
-	PendingWriteGroupedWindow PendingWriteKind = "grouped_window"
-	// PendingWriteAtomic is a caller-defined atomic batch, typically from
-	// reconciliation.
-	PendingWriteAtomic PendingWriteKind = "atomic"
-)
-
-type pendingTargetKey struct {
-	Name      string
-	Namespace string
-}
-
-// ResolvedTargetMetadata is the target-scoped planning data retained with a
-// pending write so replay does not re-fetch mutable GitTarget state.
-type ResolvedTargetMetadata struct {
-	Name             string
-	Namespace        string
-	Path             string
-	BootstrapOptions pathBootstrapOptions
-	EncryptionConfig *ResolvedEncryptionConfig
-}
-
-// PendingWrite is the unit retained until a push succeeds.
-type PendingWrite struct {
-	Kind               PendingWriteKind
-	Events             []Event
-	CommitMessage      string
-	CommitConfig       CommitConfig
-	Signer             gogit.Signer
-	GitTargetName      string
-	GitTargetNamespace string
-	Targets            map[pendingTargetKey]ResolvedTargetMetadata
-	ByteSize           int64
-}
-
-// CommitPlan is the executable plan derived from one or more retained writes.
-type CommitPlan struct {
-	Units []CommitUnit
-}
-
-// CommitMessageKind determines which message/authorship path the executor uses.
-type CommitMessageKind string
-
-const (
-	CommitMessagePerEvent CommitMessageKind = "event"
-	CommitMessageBatch    CommitMessageKind = "batch"
-	CommitMessageGrouped  CommitMessageKind = "group"
-)
-
-// CommitUnit is one locally-created commit.
-type CommitUnit struct {
-	Events        []Event
-	MessageKind   CommitMessageKind
-	CommitMessage string
-	CommitConfig  CommitConfig
-	Signer        gogit.Signer
-	GroupAuthor   string
-	Target        ResolvedTargetMetadata
-}
 
 func (w *BranchWorker) buildGroupedPendingWrite(ctx context.Context, events []Event) (*PendingWrite, error) {
 	if len(events) == 0 {
@@ -304,25 +234,6 @@ func (w *BranchWorker) buildCommitPlan(pendingWrites []PendingWrite) (CommitPlan
 	return CommitPlan{Units: units}, nil
 }
 
-func (w *BranchWorker) executeCommitPlan(ctx context.Context, repo *gogit.Repository, plan CommitPlan) (int, error) {
-	worktree, err := repo.Worktree()
-	if err != nil {
-		return 0, fmt.Errorf("failed to get worktree: %w", err)
-	}
-
-	commitsCreated := 0
-
-	for _, unit := range plan.Units {
-		created, err := w.executeCommitUnit(ctx, repo, worktree, unit)
-		if err != nil {
-			return commitsCreated, err
-		}
-		commitsCreated += created
-	}
-
-	return commitsCreated, nil
-}
-
 func (p PendingWrite) findTargetMetadata(name, namespace string) ResolvedTargetMetadata {
 	if name == "" || namespace == "" || len(p.Targets) == 0 {
 		return ResolvedTargetMetadata{}
@@ -332,134 +243,4 @@ func (p PendingWrite) findTargetMetadata(name, namespace string) ResolvedTargetM
 		return targetMetadata
 	}
 	return ResolvedTargetMetadata{}
-}
-
-func (u CommitUnit) path() string {
-	if u.Target.Path != "" {
-		return u.Target.Path
-	}
-	for _, event := range u.Events {
-		if event.Path != "" {
-			return event.Path
-		}
-	}
-	return ""
-}
-
-func (u CommitUnit) commitMetadata() (string, *gogit.CommitOptions, error) {
-	when := time.Now()
-	switch u.MessageKind {
-	case CommitMessagePerEvent:
-		if len(u.Events) != 1 {
-			return "", nil, errors.New("per-event commit unit requires exactly one event")
-		}
-		message, err := renderEventCommitMessage(u.Events[0], u.CommitConfig)
-		if err != nil {
-			return "", nil, err
-		}
-		return message, commitOptionsForEvent(u.Events[0], u.CommitConfig, u.Signer, when), nil
-	case CommitMessageBatch:
-		message, err := renderBatchCommitMessage(&WriteRequest{
-			Events:        u.Events,
-			CommitMessage: u.CommitMessage,
-			GitTargetName: u.Target.Name,
-		}, u.CommitConfig)
-		if err != nil {
-			return "", nil, err
-		}
-		return message, commitOptionsForBatch(u.CommitConfig, u.Signer, when), nil
-	case CommitMessageGrouped:
-		group := buildCommitGroupForUnit(u)
-		message, err := renderGroupCommitMessage(group, u.CommitConfig)
-		if err != nil {
-			return "", nil, err
-		}
-		return message, commitOptionsForGroup(group, u.CommitConfig, u.Signer, when), nil
-	default:
-		return "", nil, fmt.Errorf("unsupported commit message kind %q", u.MessageKind)
-	}
-}
-
-func (w *BranchWorker) executeCommitUnit(
-	ctx context.Context,
-	repo *gogit.Repository,
-	worktree *gogit.Worktree,
-	unit CommitUnit,
-) (int, error) {
-	if len(unit.Events) == 0 {
-		return 0, nil
-	}
-
-	encryptionPath := filepath.Join(worktree.Filesystem.Root(), sanitizePath(unit.path()))
-	if err := configureSecretEncryptionWriter(
-		w.contentWriter,
-		encryptionPath,
-		unit.Target.EncryptionConfig,
-	); err != nil {
-		return 0, fmt.Errorf("configure secret encryptor: %w", err)
-	}
-
-	anyChanges, err := w.applyCommitUnitEvents(ctx, repo, worktree, unit.Events)
-	if err != nil {
-		return 0, err
-	}
-	if !anyChanges {
-		return 0, nil
-	}
-
-	commitMessage, commitOptions, err := unit.commitMetadata()
-	if err != nil {
-		return 0, err
-	}
-
-	if _, err := worktree.Commit(commitMessage, commitOptions); err != nil {
-		return 0, fmt.Errorf("failed to create commit: %w", err)
-	}
-
-	log.FromContext(ctx).Info(
-		"git commit created",
-		"messageKind",
-		unit.MessageKind,
-		"events",
-		len(unit.Events),
-		"message",
-		commitMessage,
-	)
-	return 1, nil
-}
-
-func (w *BranchWorker) applyCommitUnitEvents(
-	ctx context.Context,
-	repo *gogit.Repository,
-	worktree *gogit.Worktree,
-	events []Event,
-) (bool, error) {
-	anyChanges := false
-	for _, event := range events {
-		if err := ensureBootstrapTemplateInPath(repo, sanitizePath(event.Path), event.BootstrapOptions); err != nil {
-			return false, err
-		}
-
-		changesApplied, err := applyEventToWorktree(ctx, w.contentWriter, worktree, event)
-		if err != nil {
-			return false, err
-		}
-		if changesApplied {
-			anyChanges = true
-		}
-	}
-	return anyChanges, nil
-}
-
-func buildCommitGroupForUnit(unit CommitUnit) *commitGroup {
-	group := &commitGroup{
-		Author:             unit.GroupAuthor,
-		GitTarget:          unit.Target.Name,
-		GitTargetNamespace: unit.Target.Namespace,
-		pathToEvent:        make(map[string]Event, len(unit.Events)),
-	}
-	for _, event := range unit.Events {
-		group.add(event)
-	}
-	return group
 }
